refactor: JSDocの内容を更新し、ファイル名を変更
